# Optimize power usage by adding -scan-once flag