Most of the time I do not want to scan for keyboards every 4 seconds because,
I have only one keyboard connected.

While being on the plane I've noticed high shift-shift power usage, which
seems to be linked to this scanning.

See following `powertop` results:

```
Without -scan-once:
                Usage       Events/s    Category       Description
              2.0 ms/s     235.6        Process        [PID 8593] shift-shift -match keyboard -first CAPSLOCK
            428.0 µs/s      18.6        Timer          tick_sched_timer
             19.0 ms/s       1.9        Process        [PID 3481] powertop

With -scan-once:
                Usage       Events/s    Category       Description
             14.8 ms/s       2.8        Process        [PID 3481] powertop
            116.7 µs/s       6.2        Timer          tick_sched_timer
             10.4 µs/s       6.6        kWork          flush_to_ldisc
```